feat: create board user endpoint
