+++ backend/src/modules/auth/services/register.auth.service.ts
@@ -1,3 +1,4 @@
+import { CreateBoardService } from './../../boards/interfaces/services/create.board.service.interface';
 import { UserRepositoryInterface } from './../../users/repository/user.repository.interface';
 import { GetTokenAuthService } from 'src/modules/auth/interfaces/services/get-token.auth.service.interface';
 import { BOARD_USER_NOT_FOUND, INSERT_FAILED } from 'src/libs/exceptions/messages';
@@ -8,9 +9,9 @@ import { CreateUserService } from 'src/modules/users/interfaces/services/create.
 import { TYPES } from 'src/modules/users/interfaces/types';
 import * as AUTH_TYPES from 'src/modules/auth/interfaces/types';
 import * as USER_TYPES from 'src/modules/users/interfaces/types';
+import * as BOARD_TYPES from 'src/modules/boards/interfaces/types';
 import { RegisterAuthService } from '../interfaces/services/register.auth.service.interface';
 import CreateGuestUserDto from 'src/modules/users/dto/create.guest.user.dto';
-import { BoardRoles } from 'src/libs/enum/board.roles';
 import { InjectModel } from '@nestjs/mongoose';
 import BoardUser, { BoardUserDocument } from 'src/modules/boards/entities/board.user.schema';
 import { Model } from 'mongoose';
@@ -29,7 +30,9 @@ export default class RegisterAuthServiceImpl implements RegisterAuthService {
 		@Inject(AUTH_TYPES.TYPES.services.GetTokenAuthService)
 		private getTokenAuthService: GetTokenAuthService,
 		@Inject(USER_TYPES.TYPES.repository)
-		private userRepository: UserRepositoryInterface
+		private userRepository: UserRepositoryInterface,
+		@Inject(BOARD_TYPES.TYPES.services.CreateBoardService)
+		private createBoardService: CreateBoardService
 	) {}
 
 	public async register(registrationData: CreateUserDto) {
@@ -88,28 +91,10 @@ export default class RegisterAuthServiceImpl implements RegisterAuthService {
 		const { accessToken } = await this.getTokenAuthService.getTokens(user);
 		this.userRepository.findOneByFieldAndUpdate({ _id: user }, { $set: { updatedAt: new Date() } });
 
-		await this.createGuestBoardUser(board, user);
+		await this.createBoardService.createBoardUser(board, user);
 
 		await this.sendGuestBoardUser(board, user);
 
 		return { accessToken, user };
-	}
-
-	private async createGuestBoardUser(board: string, user: string) {
-		const boardUserFound = await this.boardUserModel.findOne({ board, user });
-
-		if (boardUserFound) return;
-
-		const boardUser = {
-			role: BoardRoles.MEMBER,
-			board,
-			user,
-			votesCount: 0
-		};
-		const boardUserCreated = await this.boardUserModel.create(boardUser);
-
-		if (!boardUserCreated) throw new BadRequestException(INSERT_FAILED);
-
-		return boardUserCreated;
 	}
 }
